LPS-153063 Get the new structure for frontend tokens when the config is initilized and calculate token values once
